[ADD] Stock Transport: Update Inventory Batch: Graph, Add Inventory Batch: Gantt and add new Stock module
Add the dock stock module for add Dock into Inventory Batch form.

Update the Inventory Batch: Graph view which display total count of weigh and volume.
Add the Gantt view for Inventory Batch.

Add Stock module which is act as a bridge module. It can install the Stock Transport module. It inherit the fields of res.config.settings model and also inherit the res.config.settings view of Stock module to display the check box for installing Stock Transport Module.

+++ stock_transport/models/fleet_vehicle_model_category.py
@@ -5,8 +5,8 @@ class FleetVehicleModelCategory(models.Model):
     _name = 'fleet.vehicle.model.category'
     _inherit = 'fleet.vehicle.model.category'
 
-    max_weight = fields.Float(string="Max Weight(kg)" , default=1)
-    max_volume = fields.Float(string="Max Volume", default=1)
+    max_weight = fields.Integer(string="Max Weight(kg)" , default=1)
+    max_volume = fields.Integer(string="Max Volume", default=1)
     display_name = fields.Char()
 
     @api.depends('name')
